# backend/internal/api/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: KubeCoder API
  version: 0.1.0
  description: KubeCoder control-plane API.
servers:
  - url: /
...
paths:
...
  /api/v1/organizations/{slug}/repositories:
    parameters: [{ $ref: '#/components/parameters/OrganizationSlug' }]
    get:
      tags: [Repositories]
      operationId: listRepositories
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { name: provider, in: query, schema: { $ref: '#/components/schemas/GitProvider' } }
        - { name: configured, in: query, schema: { type: boolean } }
        - { name: agent_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Repository page, content: { application/json: { schema: { $ref: '#/components/schemas/RepositoryPage' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Repositories]
      operationId: createRepository
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RepositoryInput' } } } }
      responses:
        '201': { description: Created repository, content: { application/json: { schema: { $ref: '#/components/schemas/Repository' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/organizations/{slug}/repositories/{id}:
    parameters:
      - { $ref: '#/components/parameters/OrganizationSlug' }
      - { $ref: '#/components/parameters/ResourceID' }
    get:
      tags: [Repositories]
      operationId: getRepository
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Repository, content: { application/json: { schema: { $ref: '#/components/schemas/Repository' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Repositories]
      operationId: updateRepository
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RepositoryInput' } } } }
      responses:
        '200': { description: Updated repository, content: { application/json: { schema: { $ref: '#/components/schemas/Repository' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Repositories]
      operationId: deleteRepository
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Repository deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    Query:
      name: query
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    OrganizationSlug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GitProvider:
      type: string
      enum: [github, gitlab, bitbucket]
    PublicAgent:
      type: object
      required: [id, name, description, capabilities, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        capabilities: { type: array, items: { type: string } }
        active: { type: boolean }
...
    RepositoryInput:
      type: object
      required: [name, provider, clone_url, default_branch, include_submodules]
      properties:
        name: { type: string, maxLength: 120 }
        provider: { $ref: '#/components/schemas/GitProvider' }
        clone_url: { type: string }
        default_branch: { type: string, maxLength: 255 }
        include_submodules: { type: boolean }
        agent_id: { type: string, format: uuid, nullable: true }
      additionalProperties: false
    Repository:
      allOf:
        - { $ref: '#/components/schemas/RepositoryInput' }
        - type: object
          required: [id, organization_id, agent, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            organization_id: { type: string, format: uuid }
            agent:
              allOf: [{ $ref: '#/components/schemas/PublicAgent' }]
              nullable: true
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    RepositoryPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Repository' } }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
    Problem:
      type: object
      required: [type, title, status, detail, instance, correlation_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlation_id: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Access forbidden
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...